# backend/src/main/java/com/medibrief/backend/model/TimelineEvent.java
package com.medibrief.backend.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TimelineEvent {
    private String date;
    private String event;
}